Unify data and update request DTOs for problem settings.
Update requests may be done by creating new request DTOs or by altering the response DTO which shares the fields with the former. To support updating problem settings this way, we need to use same types for the fields of the DTOs. In this case, change the enum type to string, which is already the type used in all other DTOs.